+++ openedx_authz/settings/common.py
@@ -24,10 +24,5 @@ def plugin_settings(settings):
     settings.CASBIN_MODEL = os.path.join(
         ROOT_DIRECTORY, "engine", "config", "model.conf"
     )
-    settings.CASBIN_WATCHER_ENABLED = False
     if not hasattr(settings, "CASBIN_AUTO_LOAD_POLICY_INTERVAL"):
         settings.CASBIN_AUTO_LOAD_POLICY_INTERVAL = 5
-    # TODO: Replace with a more dynamic configuration
-    # Redis host and port are temporarily loaded here for the MVP
-    settings.REDIS_HOST = "redis"
-    settings.REDIS_PORT = 6379